# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ROMP API",
    "version": "1.0.0",
    "description": "The ROMP commerce API. Every write in the system passes through here. Generated from Zod contracts."
  },
  "paths": {
...
    "/v1/orders": {
      "post": {
        "summary": "Place the signed-in customer’s cart as an order (idempotent)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The placed order, with the UPI QR payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`INSUFFICIENT_STOCK` (Insufficient stock) — https://romp.dev/errors/insufficient-stock; `INVALID_STATE_TRANSITION` (Invalid state transition) — https://romp.dev/errors/invalid-state-transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED` (Too many requests) — https://romp.dev/errors/rate-limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Read one of the caller’s own orders",
        "responses": {
          "200": {
            "description": "The order as the customer views it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/{id}/payment-proof": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Submit a UPI payment reference (and optional proof) for an order (idempotent)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPaymentProofRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The order moved into verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitPaymentProofResponse"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`DUPLICATE_PAYMENT_REFERENCE` (Payment reference already used) — https://romp.dev/errors/duplicate-payment-reference; `RESERVATION_EXPIRED` (Reservation expired) — https://romp.dev/errors/reservation-expired; `INVALID_STATE_TRANSITION` (Invalid state transition) — https://romp.dev/errors/invalid-state-transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED` (Too many requests) — https://romp.dev/errors/rate-limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orders": {
      "get": {
        "summary": "List orders — newest first, filterable by status or fulfilment, or searched by humanId",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "fulfilmentStatus",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FulfilmentStatus"
            }
          },
          {
            "name": "humanId",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/HumanOrderId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of orders, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderListResponse"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` (Not permitted) — https://romp.dev/errors/forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cursor": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/Cursor",
        "type": "string",
        "minLength": 1,
        "maxLength": 2048
      },
...
      "HumanOrderId": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/HumanOrderId",
        "type": "string",
        "pattern": "^[A-Z]{2,8}-\\d{4,12}$"
      },
...
      "OrderStatus": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/OrderStatus",
        "type": "string",
        "enum": [
          "awaiting_payment",
          "pending_verification",
          "paid",
          "payment_rejected",
          "expired",
          "cancelled",
          "refunded"
        ]
      },
      "FulfilmentStatus": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/FulfilmentStatus",
        "type": "string",
        "enum": [
          "unfulfilled",
          "packed",
          "shipped",
          "delivered",
          "on_hold",
          "cancelled"
        ]
      },
...
      "ProblemDetails": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/ProblemDetails",
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          },
          "retryAfterSeconds": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "additionalProperties": false
      },
...
      "PlaceOrderRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/PlaceOrderRequest",
        "type": "object",
        "properties": {
          "addressId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1500
          },
          "deliverySpeed": {
            "$ref": "#/components/schemas/DeliverySpeed"
          },
          "isGift": {
            "type": "boolean"
          },
          "giftMessage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "addressId",
          "deliverySpeed",
          "isGift",
          "giftMessage"
        ],
        "additionalProperties": false
      },
      "PlaceOrderResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/PlaceOrderResponse",
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1500
          },
          "humanId": {
            "$ref": "#/components/schemas/HumanOrderId"
          },
          "qrPayload": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "amounts": {
            "type": "object",
            "properties": {
              "subtotalMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "giftWrapMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "shippingMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "taxMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "totalMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "refundedMinor": {
                "$ref": "#/components/schemas/Money"
              }
            },
            "required": [
              "subtotalMinor",
              "giftWrapMinor",
              "shippingMinor",
              "taxMinor",
              "totalMinor",
              "refundedMinor"
            ],
            "additionalProperties": false
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "required": [
          "orderId",
          "humanId",
          "qrPayload",
          "amounts",
          "status"
        ],
        "additionalProperties": false
      },
      "OrderView": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/OrderView",
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1500
          },
          "humanId": {
            "$ref": "#/components/schemas/HumanOrderId"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "fulfilment": {
            "type": "object",
            "properties": {
              "status": {
                "$ref": "#/components/schemas/FulfilmentStatus"
              },
              "carrier": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "trackingNo": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "packedAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "shippedAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "deliveredAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "holdReason": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "status",
              "carrier",
              "trackingNo",
              "packedAt",
              "shippedAt",
              "deliveredAt",
              "holdReason"
            ],
            "additionalProperties": false
          },
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 1500
                },
                "variantId": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 1500
                },
                "sku": {
                  "$ref": "#/components/schemas/Sku"
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                },
                "variantName": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                "imagePath": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "unitPriceMinor": {
                  "$ref": "#/components/schemas/Money"
                },
                "qty": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991
                },
                "lineTotalMinor": {
                  "$ref": "#/components/schemas/Money"
                }
              },
              "required": [
                "productId",
                "variantId",
                "sku",
                "name",
                "variantName",
                "imagePath",
                "unitPriceMinor",
                "qty",
                "lineTotalMinor"
              ],
              "additionalProperties": false
            }
          },
          "amounts": {
            "type": "object",
            "properties": {
              "subtotalMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "giftWrapMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "shippingMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "taxMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "totalMinor": {
                "$ref": "#/components/schemas/Money"
              },
              "refundedMinor": {
                "$ref": "#/components/schemas/Money"
              }
            },
            "required": [
              "subtotalMinor",
              "giftWrapMinor",
              "shippingMinor",
              "taxMinor",
              "totalMinor",
              "refundedMinor"
            ],
            "additionalProperties": false
          },
          "shippingAddress": {
            "type": "object",
            "properties": {
              "recipientName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 120
              },
              "line1": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "line2": {
                "anyOf": [
                  {
                    "type": "string",
                    "maxLength": 200
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "city": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "state": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "pincode": {
                "type": "string",
                "pattern": "^[1-9]\\d{5}$"
              },
              "phone": {
                "type": "string",
                "minLength": 4,
                "maxLength": 30
              }
            },
            "required": [
              "recipientName",
              "line1",
              "line2",
              "city",
              "state",
              "pincode",
              "phone"
            ],
            "additionalProperties": false
          },
          "deliverySpeed": {
            "$ref": "#/components/schemas/DeliverySpeed"
          },
          "isGift": {
            "type": "boolean"
          },
          "giftMessage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          },
          "payment": {
            "type": "object",
            "properties": {
              "method": {
                "$ref": "#/components/schemas/PaymentMethod"
              },
              "upiRef": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Utr"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "screenshotPath": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "qrPayload": {
                "type": "string",
                "minLength": 1,
                "maxLength": 2048
              },
              "submittedAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "verifiedBy": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1500
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "verifiedAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "rejectedBy": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1500
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "rejectedAt": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "rejectionReason": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "method",
              "upiRef",
              "screenshotPath",
              "qrPayload",
              "submittedAt",
              "verifiedBy",
              "verifiedAt",
              "rejectedBy",
              "rejectedAt",
              "rejectionReason"
            ],
            "additionalProperties": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "orderId",
          "humanId",
          "status",
          "fulfilment",
          "items",
          "amounts",
          "shippingAddress",
          "deliverySpeed",
          "isGift",
          "giftMessage",
          "payment",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "SubmitPaymentProofRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/SubmitPaymentProofRequest",
        "type": "object",
        "properties": {
          "upiRef": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "screenshotPath": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "upiRef",
          "screenshotPath"
        ],
        "additionalProperties": false
      },
      "SubmitPaymentProofResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/SubmitPaymentProofResponse",
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1500
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "orderId",
          "status",
          "submittedAt"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token. Admin routes additionally require a role custom claim and verify token revocation on every request."
      }
    }
  }
}